% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Automatic build of anomaly detection model}
\usage{
build(data, func = FPI, initial_support = 0.5, top_outlier_threshold = 3,
  iteration_timeout = 10)
}
\arguments{
\item{data}{data.frame with input data}

\item{func}{function name of method}

\item{initial_support}{maximum length of frequent itemsets}

\item{top_outlier_threshold}{number of top unique outliers}

\item{iteration_timeout}{timeout of one iteration}
}
\value{
model of outlier detection
}
\description{
Automatic build of anomaly detection model
}
\examples{
library("fpmoutliers")
data("iris")
model <- fpmoutliers::build(iris)

}

