% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpi.R
\name{FPI}
\alias{FPI}
\title{Frequent Pattern Isolation}
\usage{
FPI(dataFrame, minSupport = 0.3, mlen = 0)
}
\arguments{
\item{dataFrame}{data.frame with input data}

\item{minSupport}{minimum support for FPM}

\item{mlen}{maximum length of frequent itemsets}
}
\value{
vector with outlier scores
}
\description{
Frequent Pattern Isolation
}
\examples{
library("fpmoutliers")
dataFrame <- read.csv(
     system.file("extdata", "fp-outlier-customer-data.csv", package = "fpmoutliers"))
model <- FPI(dataFrame, minSupport = 0.001)
}

