% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpi2.R
\name{FPI2}
\alias{FPI2}
\title{Frequent Pattern Isolation}
\usage{
FPI2(dataFrame, minSupport = 0.3, mlen = 0)
}
\arguments{
\item{dataFrame}{data.frame with input data}

\item{minSupport}{minimum support for FPM}

\item{mlen}{maximum length of frequent itemsets}
}
\value{
vector with outlier scores
}
\description{
Frequent Pattern Isolation
}

