% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpof.R
\name{FPOF}
\alias{FPOF}
\title{Frequent Pattern Outlier Factor}
\usage{
FPOF(dataFrame, minSupport = 0.3, mlen = 0)
}
\arguments{
\item{dataFrame}{data.frame with input data}

\item{minSupport}{minimum support for FPM}

\item{mlen}{maximum length of frequent itemsets}
}
\value{
vector with outlier scores
}
\description{
He, Z., Xu, X., Huang, J. Z., Deng, S.: FP-Outlier: Frequent Pattern Based Outlier Detection. Computer Science and Information Systems, Vol. 2, No. 1, 103-118. (2005)
}

