% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpof.R
\name{FPOFcontradictness}
\alias{FPOFcontradictness}
\title{Frequent Pattern Outlier Factor}
\usage{
FPOFcontradictness(dataFrame, anIndex, minSupport = 0.3, mlen = 0, k = 10)
}
\arguments{
\item{dataFrame}{data.frame with input data}

\item{anIndex}{anomaly index}

\item{minSupport}{minimum support for FPM}

\item{mlen}{maximum length of frequent itemsets}

\item{k}{top-k contradictness}
}
\value{
vector with outlier scores
}
\description{
Frequent Pattern Outlier Factor
}
