% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmml.R
\name{generatePMML}
\alias{generatePMML}
\title{PMML conversion}
\usage{
generatePMML(model, dataFrame = NULL, topN = NULL)
}
\arguments{
\item{model}{outlier model}

\item{dataFrame}{frame for labeling}

\item{topN}{limit number of outliers in the output}
}
\value{
pmml model
}
\description{
PMML conversion
}
\examples{
library("fpmoutliers")
dataFrame <- read.csv(
     system.file("extdata", "fp-outlier-customer-data.csv", package = "fpmoutliers"))
model <- FPI(dataFrame, minSupport = 0.001)
#generatePMML(model, dataFrame)
}
