% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfpof.R
\name{LFPOF}
\alias{LFPOF}
\title{LFPOF}
\usage{
LFPOF(data, minSupport = 0.3, mlen = 0)
}
\arguments{
\item{data}{\code{data.frame} or \code{transactions} from \code{arules} with input data}

\item{minSupport}{minimum support for FPM}

\item{mlen}{maximum length of frequent itemsets}
}
\value{
vector with outlier scores
}
\description{
W. Zhang, J. Wu and J. Yu, "An Improved Method of Outlier Detection Based on Frequent Pattern," Information Engineering (ICIE), 2010 WASE International Conference on, Beidaihe, Hebei, 2010, pp. 3-6.
}

