% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfpof.R
\name{MFPOF}
\alias{MFPOF}
\title{MFPOF}
\usage{
MFPOF(dataFrame, minSupport = 0.3, mlen = 0)
}
\arguments{
\item{dataFrame}{data.frame with input data}

\item{minSupport}{minimum support for FPM}

\item{mlen}{maximum length of frequent itemsets}
}
\value{
vector with outlier scores
}
\description{
Feng Lin, Wang Le, Jin Bo - Research on Maximal Frequent Pattern Outlier Factor for Online HighDimensional Time-Series Outlier Detection. Journal of Convergence Information Technology 5(10):66-71. December 2010.
}

