% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmml.R
\name{parsePMML}
\alias{parsePMML}
\title{PMML parser}
\usage{
parsePMML(fileName)
}
\arguments{
\item{fileName}{xml file name}
}
\value{
list model
}
\description{
The parser parses the proposed PMML for the outlier detection model and build its object representation.
}
\examples{
\dontrun{
library("fpmoutliers")
model <- parsePMML("od-pmml.xml")
}
}
