% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcfpof.R
\name{WCFPOF}
\alias{WCFPOF}
\title{WCFPOF}
\usage{
WCFPOF(dataFrame, minSupport = 0.3, mlen = 0)
}
\arguments{
\item{dataFrame}{data.frame with input data}

\item{minSupport}{minimum support for FPM}

\item{mlen}{maximum length of frequent itemsets}
}
\value{
vector with outlier scores
}
\description{
Jiadong Ren, Qunhui Wu,  Jiadong Ren, Changzhen Hu, Kunsheng Wang - An Approach for Analyzing Infrequent Software Faults Based on Outlier Detection
}

